Create one-line-snake-unminified.py
So turns out the unminified version is ± 160 lines of code.
It would probably be a different story if it was some clean code without hacks to make it run in one single instruction. It would be lot more boring too.
The "unminified" version only exist to comment/partially-explain the black magic behind the minified version.